Refactoring Extract Class
