+++ propagator/ottrace/lib/opentelemetry-propagator-ottrace.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+#
+# Copyright The OpenTelemetry Authors
+#
+# SPDX-License-Identifier: Apache-2.0
+
+require 'opentelemetry-api'
+require_relative './opentelemetry/propagator/ottrace'
